pyproject.toml
[project]
name = "1st-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.1.0",
    "pyautogen>=0.9",
    "pydantic-ai>=0.4.2",
    "python-dotenv>=1.1.1",
]
